refactor: reorganize directory structure of HTML-related files
